add dynamic precison detect
