+++ point/decode_test.go
@@ -87,6 +87,85 @@ func TestTimeRound(t *T.T) {
 	})
 }
 
+func TestDynamicPrecision(t *T.T) {
+	pts := []*Point{
+		func() *Point {
+			var kvs KVs
+			kvs = kvs.AddV2("f1", 123, true)
+			return NewPointV2("p1", kvs, WithTimestamp(1716536956))
+		}(),
+
+		func() *Point {
+			var kvs KVs
+			kvs = kvs.AddV2("f1", 123, true)
+			return NewPointV2("p1", kvs, WithTimestamp(1716536956000))
+		}(),
+
+		func() *Point {
+			var kvs KVs
+			kvs = kvs.AddV2("f1", 123, true)
+			return NewPointV2("p1", kvs, WithTimestamp(1716536956000000))
+		}(),
+
+		func() *Point {
+			var kvs KVs
+			kvs = kvs.AddV2("f1", 123, true)
+			return NewPointV2("p1", kvs, WithTimestamp(1716536956000000000))
+		}(),
+	}
+
+	cases := []struct {
+		name string
+		e    Encoding
+	}{
+		{
+			"line-protocol",
+			LineProtocol,
+		},
+
+		//{
+		//	"json",
+		//	JSON,
+		//},
+
+		//{
+		//	"pbjson",
+		//	PBJSON,
+		//},
+
+		{
+			"pb",
+			Protobuf,
+		},
+	}
+
+	for _, tc := range cases {
+		t.Run(tc.name, func(t *T.T) {
+			enc := GetEncoder(WithEncEncoding(tc.e))
+			defer PutEncoder(enc)
+
+			enc.EncodeV2(pts)
+			buf := make([]byte, 1<<20) // large buffer
+			var (
+				encBuf []byte
+				ok     bool
+			)
+
+			encBuf, ok = enc.Next(buf) // encode once we should get all the buffer
+			assert.True(t, ok, "enc last error: %s", enc.LastErr())
+
+			dec := GetDecoder(WithDecEncoding(tc.e))
+			defer PutDecoder(dec)
+
+			newPts, err := dec.Decode(encBuf, WithPrecision(PrecDyn))
+			assert.NoError(t, err)
+			for _, pt := range newPts {
+				assert.Equal(t, int64(1716536956000000000), pt.pt.Time)
+			}
+		})
+	}
+}
+
 func TestDecode(t *T.T) {
 	var fnCalled int
