+++ point/encode.go
@@ -13,81 +13,6 @@ import (
 	sync "sync"
 )
 
-type Encoding int
-
-const (
-	encProtobuf      = "protobuf"
-	encProtobufAlias = "v2"
-	encJSON          = "json"
-
-	encLineprotocolAlias = "v1"
-	encLineprotocol      = "line-protocol"
-
-	contentTypeJSON      = "application/json"
-	contentTypeProtobuf  = "application/protobuf; proto=com.guance.Point"
-	contentTypeLineproto = "application/line-protocol"
-)
-
-// EncodingStr convert encoding-string in configure file to
-// encoding enum.
-//
-// Here v1/v2 are alias for lineprotocol and protobuf, this makes
-// people easy to switch between lineprotocol and protobuf. For
-// json, you should not configure json encoding in production
-// environments(json do not classify int and float).
-func EncodingStr(s string) Encoding {
-	switch strings.ToLower(s) {
-	case encProtobuf, encProtobufAlias:
-		return Protobuf
-	case encJSON:
-		return JSON
-	case encLineprotocol,
-		encLineprotocolAlias:
-		return LineProtocol
-	default:
-		return LineProtocol
-	}
-}
-
-func HTTPContentType(ct string) Encoding {
-	switch ct {
-	case contentTypeJSON:
-		return JSON
-	case contentTypeProtobuf:
-		return Protobuf
-	case contentTypeLineproto:
-		return LineProtocol
-	default: // default use line-protocol to be compatible with lagacy code
-		return LineProtocol
-	}
-}
-
-func (e Encoding) HTTPContentType() string {
-	switch e {
-	case JSON:
-		return contentTypeJSON
-	case Protobuf:
-		return contentTypeProtobuf
-	case LineProtocol:
-		return contentTypeLineproto
-	default: // default use line-protocol to be compatible with lagacy code
-		return contentTypeLineproto
-	}
-}
-
-func (e Encoding) String() string {
-	switch e {
-	case JSON:
-		return encJSON
-	case Protobuf:
-		return encProtobuf
-	case LineProtocol:
-		return encLineprotocol
-	default: // default use line-protocol to be compatible with lagacy code
-		return encLineprotocol
-	}
-}
-
 // EncodeFn used to iterate on []*Point payload, if error returned, the iterate terminated.
 type EncodeFn func(batchSize int, payload []byte) error
 
@@ -297,151 +222,7 @@ func (e *Encoder) Encode(pts []*Point) ([][]byte, error) {
 	return e.doEncode(pts)
 }
 
-func (e *Encoder) EncodeV2(pts []*Point) {
-	e.pts = pts
-}
-
-func (e *Encoder) doEncodeProtobuf(buf []byte) ([]byte, bool) {
-	var (
-		curSize,
-		pbptsSize int
-		trimmed = 1
-	)
-
-	for _, pt := range e.pts[e.lastPtsIdx:] {
-		if pt == nil {
-			continue
-		}
-
-		curSize += pt.Size()
-
-		// e.pbpts size larger than buf, we must trim some of points
-		// until size fit ok or MarshalTo will panic.
-		if curSize >= len(buf) {
-			if len(e.pbpts.Arr) <= 1 { // nothing to trim
-				e.lastErr = errTooSmallBuffer
-				return nil, false
-			}
-
-			for {
-				if pbptsSize = e.pbpts.Size(); pbptsSize > len(buf) {
-					e.pbpts.Arr = e.pbpts.Arr[:len(e.pbpts.Arr)-trimmed]
-					e.lastPtsIdx -= trimmed
-					trimmed *= 2
-				} else {
-					goto __doEncode
-				}
-			}
-		} else {
-			e.pbpts.Arr = append(e.pbpts.Arr, pt.pt)
-			e.lastPtsIdx++
-		}
-	}
-
-__doEncode:
-	e.trimmed = trimmed
-
-	if len(e.pbpts.Arr) == 0 {
-		return nil, false
-	}
-
-	defer func() {
-		e.pbpts.Arr = e.pbpts.Arr[:0]
-	}()
-
-	if n, err := e.pbpts.MarshalTo(buf); err != nil {
-		e.lastErr = err
-		return nil, false
-	} else {
-		if e.fn != nil {
-			if err := e.fn(len(e.pbpts.Arr), buf[:n]); err != nil {
-				e.lastErr = err
-				return nil, false
-			}
-		}
-
-		e.parts++
-		return buf[:n], true
-	}
-}
-
 var errTooSmallBuffer = errors.New("too small buffer")
-
-func (e *Encoder) doEncodeLineProtocol(buf []byte) ([]byte, bool) {
-	curSize := 0
-	npts := 0
-
-	for _, pt := range e.pts[e.lastPtsIdx:] {
-		if pt == nil {
-			continue
-		}
-
-		lppt, err := pt.LPPoint()
-		if err != nil {
-			e.lastErr = err
-			continue
-		}
-
-		ptsize := lppt.StringSize()
-
-		if curSize+ptsize+1 > len(buf) { // extra +1 used to store the last '\n'
-			if curSize == 0 { // nothing added
-				e.lastErr = errTooSmallBuffer
-				return nil, false
-			}
-
-			if e.fn != nil {
-				if err := e.fn(npts, buf[:curSize]); err != nil {
-					e.lastErr = err
-					return nil, false
-				}
-			}
-			e.parts++
-			return buf[:curSize], true
-		} else {
-			e.lpPointBuf = lppt.AppendString(e.lpPointBuf)
-
-			copy(buf[curSize:], e.lpPointBuf[:ptsize])
-
-			// Always add '\n' to the end of current point, this may
-			// cause a _unneeded_ '\n' to the end of buf, it's ok for
-			// line-protocol parsing.
-			buf[curSize+ptsize] = '\n'
-			curSize += (ptsize + 1)
-
-			// clean buffer, next time AppendString() append from byte 0
-			e.lpPointBuf = e.lpPointBuf[:0]
-			e.lastPtsIdx++
-			npts++
-		}
-	}
-
-	if curSize > 0 {
-		e.parts++
-		if e.fn != nil { // NOTE: encode callback error will terminate encode
-			if err := e.fn(npts, buf[:curSize]); err != nil {
-				e.lastErr = err
-				return nil, false
-			}
-		}
-		return buf[:curSize], true
-	} else {
-		return nil, false
-	}
-}
-
-func (e *Encoder) Next(buf []byte) ([]byte, bool) {
-	switch e.enc {
-	case Protobuf:
-		return e.doEncodeProtobuf(buf)
-	case LineProtocol:
-		return e.doEncodeLineProtocol(buf)
-	case JSON:
-		return nil, false
-	default: // TODO: json
-		return nil, false
-	}
-}
 
 func (e *Encoder) LastErr() error {
 	return e.lastErr
